utils/testing: adopt `parametrize` helper
