+++ plover_build_utils/testing/__init__.py
@@ -1,3 +1,4 @@
 from .blackbox import blackbox_test
 from .output import CaptureOutput
+from .parametrize import parametrize
 from .steno import steno_to_stroke
